// projects/ui-builder-component-lib/src/lib/enums/form-error-messages.enum.ts
export enum PasswordErrorMessages {
  minimum = 'VALIDATIONS.PASSWORD_MINIMUM',
  uppercase = 'VALIDATIONS.PASSWORD_UPPERCASE',
  lowercase = 'VALIDATIONS.PASSWORD_LOWERCASE',
  number = 'VALIDATIONS.PASSWORD_NUMBER',
  specialCharacter = 'VALIDATIONS.PASSWORD_SPECIAL_CHARACTER',
  prefix = 'VALIDATIONS.PASSWORD_ERROR_PREFIX',
}